refactor: field-note/_components 하위 ui 컴포넌트 공통 컴포넌트로 이동 / 공통 컴포넌트 디렉터리 레벨 구조 변경
