+++ frontend/eodi-web/app/field-notes/new/_components/field/DetailRecordFields.jsx
