+++ frontend/eodi-web/app/field-notes/new/_components/field/DetailStatusFields.jsx
