+++ frontend/eodi-web/app/field-notes/new/_components/field/RegionInfo.jsx
