+++ frontend/eodi-web/app/search/region-select-step/_components/MultiButtonSelectGrid.jsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ToggleButton from "@/components/ui/input/ToggleButton";
+import ToggleButton from "@/components/ui/ToggleButton";
 
 export default function MultiButtonSelectGrid({ list, selected, onSelect, placeholder }) {
   return (
