+++ frontend/eodi-web/app/search/step2/DealListPage.jsx
@@ -1,9 +1,9 @@
 "use client";
 
-import PageHeader from "@/components/ui/header/PageHeader";
+import PageHeader from "@/components/layout/PageHeader";
 import GridGroup from "@/app/search/_components/GridGroup";
 import InnerNavContainer from "@/components/layout/InnerNavContainer";
-import FloatingContainer from "@/components/ui/container/floating/FloatingContainer";
+import FloatingContainer from "@/components/ui/container/FloatingContainer";
 import { SlidersHorizontal } from "lucide-react";
 
 import RegionFilterBar from "@/app/search/step2/_components/RegionFilterBar";
@@ -11,7 +11,7 @@ import DealResultSection from "@/app/search/step2/_components/DealResultSection"
 import FloatingFilterCardContents from "@/app/search/step2/_components/FloatingFilterCardContents";
 import { useDealListPageController } from "@/app/search/step2/_hooks/useDealListPageController";
 import HousingTypeFilterBar from "@/app/search/step2/_components/HousingTypeFilterBar";
-import RadioGroup from "@/components/ui/input/RadioGroup";
+import RadioGroup from "@/components/ui/RadioGroup";
 
 export default function DealListPage() {
   const controller = useDealListPageController();
