+++ frontend/eodi-web/components/ui/DetailSectionToggle.jsx
