+++ frontend/eodi-web/components/ui/DiscreteSliderInput.jsx
