+++ frontend/eodi-web/components/ui/SliderInput.jsx
