+++ frontend/eodi-web/components/ui/container/FloatingCard.jsx
